feat: 서버 이벤트시 레디스 캐싱 및 카프카에 이벤트 발행하도록 로직 구현, 서버 조회시 없으면 캐싱하는 로직 구현

+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/kafka/dto/serverEvent/ServerChannelEventDto.java
@@ -17,7 +17,7 @@ public class ServerChannelEventDto extends ServerEventDto {
 
 	String channelType; // CHAT, VOICE
 
-	Long order; // Channel 순서, DELETE 시에는 null
+	int order; // Channel 순서, DELETE 시에는 null
 
 	String status; // CREATE, UPDATE, DELETE
